% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_names}
\alias{clean_names}
\title{Clean a vector or species names}
\usage{
clean_names(taxa_names)
}
\arguments{
\item{taxa_names}{A vector of species names.}
}
\value{
A vector of clean species names.
}
\description{
Remove any "subsp." or similar from name
}
