% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_with_retry}
\alias{download_with_retry}
\title{Function to download large Excel files from MITECO}
\usage{
download_with_retry(url, destfile, tries = 8)
}
\arguments{
\item{url}{URL to download.}

\item{destfile}{Path to downloaded file.}

\item{tries}{Number of download tries.}
}
\value{
A file
}
\description{
Function to download large Excel files from MITECO
}
