% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eidos_clean_names}
\alias{eidos_clean_names}
\title{Clean a single species names}
\usage{
eidos_clean_names(taxa_names)
}
\arguments{
\item{taxa_names}{A single species names.}
}
\value{
A single clean species names.
}
\description{
Removes any "subsp.", "var.", "f.", years and authorship similar from a species name.
}
