% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eidos_clean_whitespaces}
\alias{eidos_clean_whitespaces}
\title{Remove whitespaces}
\usage{
eidos_clean_whitespaces(x)
}
\arguments{
\item{x}{A vector of strings.}
}
\value{
A vector of strings with only " " whitespaces
}
\description{
Remove whitespaces
}
