% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_example_data.R
\docType{data}
\name{eidos_example_data}
\alias{eidos_example_data}
\title{Sample species data}
\format{
\subsection{\code{eidos_example_data}}{

A data frame with 500 rows and 4 columns:
\describe{
\item{taxon}{Full taxon name}
\item{genus}{Genus}
\item{species}{Specific epithet}
\item{subspecies}{Subspecific epithet}
}
}
}
\source{
\url{https://biotrend.es/}
}
\usage{
eidos_example_data
}
\description{
A sample of species data included in the Biotrend project
}
\keyword{datasets}
