% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_fuzzy_names.R
\name{eidos_fuzzy_names}
\alias{eidos_fuzzy_names}
\title{Fuzzy matching with the Spanish checklist of Wildlife Species}
\usage{
eidos_fuzzy_names(
  taxa_list,
  checklist = eidos_clean_checklist(),
  maxdist = 0.1,
  method = "jw",
  mode = "left",
  distance_col = "dist",
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL
)
}
\arguments{
\item{taxa_list}{A data frame or vector with taxonomic information to match with the Checklist. Must include at least the genus
and species columns. Subspecies is optional. Additional taxonomic levels from
kingdom to family can be supplied to filter any possible conflicts. These can be provided as columns with
the corresponding name in the data frame or as function arguments if the supplied
data was a vector of names.
Taxonomic authorities are not supported yet.}

\item{checklist}{A data frame. The result of running the eidos_clean_checklist() function.}

\item{maxdist}{A number. Maximum dissimilarity distance between taxa names to match. Default for "jw" is 0.1 (10\%)}

\item{method}{A string. Method to calculate the distance between names inherited from fuzzyjoin::stringdist_join. One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw" or "soundex". Default is "jw".}

\item{mode}{A string. Type of join, one of "inner", "left", "right", "full", "semi" or "anti" inherited from fuzzyjoin::stringdist_join.}

\item{distance_col}{A string. Name of the column to display the dissimilarity distance between matched names. Set NULL to omit the column.}

\item{kingdom}{A vector of strings of length = taxa_list with the kingdom of each supplied taxon.}

\item{phylum}{A vector of strings of length = taxa_list with the phylum of each supplied taxon.}

\item{class}{A vector of strings of length = taxa_list with the class of each supplied taxon.}

\item{order}{A vector of strings of length = taxa_list with the order of each supplied taxon.}

\item{family}{A vector of strings of length = taxa_list with the family of each supplied taxon.}
}
\value{
A data frame
}
\description{
Matches a data frame or vector with species names to any names in the Spanish checklist of Wildlife Species (\emph{Lista Patrón de Especies Silvestres}). Uses fuzzy matching to allow spelling errors. In addition, you can filter by higher taxonomic levels to refine the match.
}
\examples{
\donttest{
checklist = eidos_clean_checklist()
taxa_vector = c("Bordere chouardii", "Alts cisternasii")
matched_names = eidos_fuzzy_names(taxa_list = taxa_vector, checklist = checklist)

# Some names have conflicts when using fuzzy matching.
# This returns two matches for two different genera,
# Lanius (our species of interest, a bird) and Lasius (an ant).
taxa_df = data.frame(genus = "Lanius", species = "meridionalis")
matched_names = eidos_fuzzy_names(taxa_list = taxa_df, checklist = checklist)

# We can refine the search by including higher taxonomic levels:
taxa_df = data.frame(class = "Aves", genus = "Lanius", species = "meridionalis")
refined_matched_names = eidos_fuzzy_names(taxa_list = taxa_df, checklist = checklist)

# Or using a vector instead
taxa_vector = c("Lanius meridionalis")
class_vector = c("Aves")
refined_matched_names = eidos_fuzzy_names(taxa_list = taxa_vector,
                                          checklist = checklist,
                                          class = class_vector)
}

}
