% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_legal_status_by_id.R
\name{eidos_legal_status_by_id}
\alias{eidos_legal_status_by_id}
\title{Get legal status from EIDOS}
\usage{
eidos_legal_status_by_id(taxon_id)
}
\arguments{
\item{taxon_id}{An integer. A valid taxon identifier from EIDOS.}
}
\value{
A data frame.
}
\description{
Retrieves the legal status of a given taxon identifier from the EIDOS database.
Identifiers can be retrieved using the
eidos_fuzzy_names() or eidos_taxon_by_name() functions.
}
\examples{
eidos_legal_status_by_id(taxon_id = 1)
}
