% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_taxon_by_id.R
\name{eidos_taxon_by_id}
\alias{eidos_taxon_by_id}
\title{Get taxonomic infomration from ID}
\usage{
eidos_taxon_by_id(taxon_id)
}
\arguments{
\item{taxon_id}{An integer. A valid taxon identifier from EIDOS}
}
\value{
A data frame with the taxonomic information for all taxa related to that ID
}
\description{
Get taxonomic infomration from ID
}
\examples{
eidos_taxon_by_id(taxon_id = 1)
}
