% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_authorities}
\alias{get_authorities}
\title{Get scientific author from EIDOS}
\usage{
get_authorities(taxa_auth)
}
\arguments{
\item{taxa_auth}{a vector of species names}
}
\value{
DF with supplied name and authorship in eidos. Only works for vectors of species names
}
\description{
Get scientific author from EIDOS
}
