% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_clean_checklist.R
\name{eidos_clean_checklist}
\alias{eidos_clean_checklist}
\title{Get the Spanish checklist with synonyms}
\usage{
eidos_clean_checklist()
}
\value{
A data frame with the Spanish checklist with synonyms in a long format ready to use in other functions in the \emph{eidos_api} package.
}
\description{
Get the Spanish checklist with synonyms
}
\examples{
\dontrun{
eidos_clean_checklist()
}

}
