% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_conservation_by_id.R
\name{eidos_conservation_by_id}
\alias{eidos_conservation_by_id}
\title{Get the conservation information from EIDOS for a given taxa}
\usage{
eidos_conservation_by_id(taxon_id, taxo_info = F, latest = F)
}
\arguments{
\item{taxon_id}{An integer. A taxon identifier from EIDOS.}

\item{taxo_info}{Booelan. Return the taxonomic information of the taxon}

\item{latest}{Booelan. Return the information form only the latest assessment}
}
\value{
A data frame with the conservation information of each taxa.
}
\description{
Get the conservation information from EIDOS for a given taxa
}
\examples{
eidos_conservation_by_id(taxon_id = 1, taxon_info = F, latest = F)
}
