% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_legal_status_by_id.R
\name{eidos_legal_status_by_id}
\alias{eidos_legal_status_by_id}
\title{Get legal status from EIDOS}
\usage{
eidos_legal_status_by_id(taxon_id)
}
\arguments{
\item{taxon_id}{An integer. A valid taxon identifier from EIDOS.}
}
\value{
A data frame.
}
\description{
Retrieves the legal status of a given taxon identifier from thge EIDOS database.
}
\examples{
eidos_legal_status_by_id(taxon_id = 1)
}
