% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_tables.R
\name{eidos_tables}
\alias{eidos_tables}
\title{Retrieve one of the tables related to the Spanish checklist}
\usage{
eidos_tables(
  eidos_table = c("comunidades_autonomas", "listapatronespecie_codigos",
    "listapatronespecie", "componente_tema", "regbiogeograf_termar",
    "listapatronespecie_sinonimos", "pais", "norma", "provincias",
    "listapatronespecie_normas")
)
}
\arguments{
\item{table}{Name of the table to query. The function if case insensitive and admits whitespace, but the parameter must match one of "comunidades_autonomas", "listapatronespecie_codigos",
"listapatronespecie", "componente_tema", "regbiogeograf_termar", "listapatronespecie_sinonimos", "pais",
"norma", "provincias" or "listapatronespecie_normas".}
}
\value{
The desired table from the EIDOS API
}
\description{
Retrieve one of the tables related to the Spanish checklist
}
\examples{
eidos_tables(eidos_table = "listapatronespecie")
}
