% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_tables.R
\name{eidos_tables}
\alias{eidos_tables}
\title{Retrieve one of the tables related to the Spanish Checklist of Wildlie Species}
\usage{
eidos_tables(
  eidos_table = c("comunidades_autonomas", "listapatronespecie_codigos",
    "listapatronespecie", "componente_tema", "regbiogeograf_termar",
    "listapatronespecie_sinonimos", "pais", "norma", "provincias",
    "listapatronespecie_normas")
)
}
\arguments{
\item{eidos_table}{Name of the table to query. The function is case insensitive, admits whitespaces and has partial matching for arguments, but these must match one of "comunidades_autonomas", "listapatronespecie_codigos",
"listapatronespecie", "componente_tema", "regbiogeograf_termar", "listapatronespecie_sinonimos", "pais",
"norma", "provincias" or "listapatronespecie_normas".}
}
\value{
The desired table from the EIDOS API
}
\description{
The tables are downloaded verbatim from the API. This means that column names may not match those in other tables from EIDOS.
}
\examples{
eidos_tables(eidos_table = "listapatronespecie")
}
