% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_fuzzy_names.R
\name{eidos_tables}
\alias{eidos_tables}
\title{Retrieve one of the tables related to the Spanish checklist}
\usage{
eidos_tables(eidos_table = "listapatronespecie")
}
\arguments{
\item{table}{Name of the table to query}
}
\value{
The desired table from the EIDOS API
}
\description{
Retrieve one of the tables related to the Spanish checklist
}
\examples{
eidos_tables(eidos_table = "listapatronespecie")
}
