% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidos_taxon_by_id.R
\name{eidos_taxon_by_id}
\alias{eidos_taxon_by_id}
\title{Get taxonomic information from EIDOS}
\usage{
eidos_taxon_by_id(taxon_id)
}
\arguments{
\item{taxon_id}{An integer or vector of integers. Valid taxon identifiers from EIDOS.}
}
\value{
A data frame.
}
\description{
Retrieves the taxonomic information for a given taxon identifier from the EIDOS database.
Identifiers can be retrieved using the
eidos_fuzzy_names() or eidos_taxon_by_name() functions.
}
\examples{
eidos_taxon_by_id(taxon_id = 1)
}
