% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidosapi-package.R
\docType{package}
\name{eidosapi-package}
\alias{eidosapi}
\alias{eidosapi-package}
\title{The eidosapi R package}
\description{
Provides access to 'EIDOS', the taxonomic information service from the Spanish Inventory of Natural Patrimony and Biodiversity directly from R.
}
\details{
This is an 'API' to access taxonomic information from the Spanish Inventory of Natural Patrimony and Biodiversity.
For a complete list of functions use \code{library(help = "eidosapi")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmirceb/eidosapi}
  \item Report bugs at \url{https://github.com/hmirceb/eidosapi/issues}
}

}
\author{
\strong{Maintainer}: Héctor Miranda-Cebrián \email{hectorm94@gmail.com} (\href{https://orcid.org/0000-0002-0340-8373}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Héctor Miranda-Cebrián \email{hectorm94@gmail.com} (\href{https://orcid.org/0000-0002-0340-8373}{ORCID}) [copyright holder]
}

}
\keyword{internal}
